arael-sketch-backend: dimensions carry a permanent id
Dimension gains did, minted with constraint naming (max existing + 1,
zero ids from older files fixed up on the same path). Selection and
the dimension actions (Update/Remove/Move) carry the did and resolve
it at apply time, so removing one dimension no longer shifts the
identity of the rest: the GUI multi-delete deletes what was selected,
a stale did is a no-op instead of a wrong-target delete, and the
dimension edit overlay resolves its did per frame -- after an undo it
closes gracefully instead of indexing a shifted slot.

+++ arael-sketch-backend/src/ids.rs
@@ -18,7 +18,8 @@ pub enum Selection {
     ArcStart(Ref<Arc>),
     ArcEnd(Ref<Arc>),
     Constraint(ConstraintId),
-    Dimension(usize),
+    /// By permanent dimension id (Dimension::did), not Vec index.
+    Dimension(u32),
 }
 
 #[derive(Clone, Copy, PartialEq, serde::Serialize, serde::Deserialize)]
